fix(KB-280): fix History tab data fetching and refactor page.tsx
- Fetch missed_discovery and ingestion_queue separately, merge in JS
- FK join wasn't working reliably with client-side Supabase
- Extract useAddArticle hook to reduce page.tsx (501 → 314 lines)
- Add PR state check rule to .windsurfrules
