+++ admin-next/src/app/(dashboard)/add/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useAddArticle } from './useAddArticle';
